# tutorial/docs/openapi.json
{
  "components": {
    "schemas": {
...
      "CustodyReportResponse": {
        "additionalProperties": false,
        "description": "Per-evidence custody keyed by evidence id.",
        "properties": {
          "chains": {
            "additionalProperties": {
              "items": {},
              "type": "array"
            },
            "title": "Chains",
            "type": "object"
          },
          "incident_id": {
            "title": "Incident Id",
            "type": "string"
          }
        },
        "required": [
          "incident_id",
          "chains"
        ],
        "title": "CustodyReportResponse",
        "type": "object"
      },
      "EvidenceResponse": {
        "description": "Evidence metadata.",
        "properties": {
          "collected_by": {
            "title": "Collected By",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "custody_chain": {
            "items": {},
            "title": "Custody Chain",
            "type": "array"
          },
          "evidence_type": {
            "title": "Evidence Type",
            "type": "string"
          },
          "file_path": {
            "title": "File Path",
            "type": "string"
          },
          "file_size_bytes": {
            "title": "File Size Bytes",
            "type": "integer"
          },
          "hash_sha256": {
            "title": "Hash Sha256",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "incident_id": {
            "format": "uuid",
            "title": "Incident Id",
            "type": "string"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "storage_location": {
            "title": "Storage Location",
            "type": "string"
          },
          "verified_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verified At"
          }
        },
        "required": [
          "id",
          "incident_id",
          "evidence_type",
          "file_path",
          "hash_sha256",
          "file_size_bytes",
          "metadata",
          "collected_by",
          "custody_chain",
          "storage_location",
          "created_at",
          "verified_at"
        ],
        "title": "EvidenceResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InvestigationStepResponse": {
        "additionalProperties": false,
        "description": "Single investigation step.",
        "properties": {
          "action_taken": {
            "title": "Action Taken",
            "type": "string"
          },
          "agent_name": {
            "title": "Agent Name",
            "type": "string"
          },
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "correction_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Correction Reason"
          },
          "execution_time_ms": {
            "title": "Execution Time Ms",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "incident_id": {
            "format": "uuid",
            "title": "Incident Id",
            "type": "string"
          },
          "interpretation": {
            "title": "Interpretation",
            "type": "string"
          },
          "is_self_correction": {
            "title": "Is Self Correction",
            "type": "boolean"
          },
          "raw_output": {
            "title": "Raw Output",
            "type": "string"
          },
          "timestamp": {
            "format": "date-time",
            "title": "Timestamp",
            "type": "string"
          },
          "tool_used": {
            "title": "Tool Used",
            "type": "string"
          }
        },
        "required": [
          "id",
          "incident_id",
          "agent_name",
          "action_taken",
          "tool_used",
          "raw_output",
          "interpretation",
          "confidence",
          "timestamp",
          "is_self_correction",
          "correction_reason",
          "execution_time_ms"
        ],
        "title": "InvestigationStepResponse",
        "type": "object"
      },
      "KnowledgeEdgeResponse": {
        "additionalProperties": false,
        "description": "Knowledge graph edge.",
        "properties": {
          "evidence_count": {
            "title": "Evidence Count",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "relation_type": {
            "title": "Relation Type",
            "type": "string"
          },
          "source_id": {
            "title": "Source Id",
            "type": "string"
          },
          "target_id": {
            "title": "Target Id",
            "type": "string"
          },
          "weight": {
            "title": "Weight",
            "type": "number"
          }
        },
        "required": [
          "id",
          "source_id",
          "target_id",
          "relation_type",
          "weight",
          "evidence_count"
        ],
        "title": "KnowledgeEdgeResponse",
        "type": "object"
      },
      "KnowledgeGraphResponse": {
        "additionalProperties": false,
        "description": "Full graph snapshot.",
        "properties": {
          "edges": {
            "items": {
              "$ref": "#/components/schemas/KnowledgeEdgeResponse"
            },
            "title": "Edges",
            "type": "array"
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/KnowledgeNodeResponse"
            },
            "title": "Nodes",
            "type": "array"
          }
        },
        "required": [
          "nodes",
          "edges"
        ],
        "title": "KnowledgeGraphResponse",
        "type": "object"
      },
      "KnowledgeNodeResponse": {
        "additionalProperties": false,
        "description": "Knowledge graph node.",
        "properties": {
          "category": {
            "title": "Category",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "difficulty": {
            "title": "Difficulty",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "incidents_demonstrating": {
            "items": {},
            "title": "Incidents Demonstrating",
            "type": "array"
          },
          "lessons_teaching": {
            "items": {},
            "title": "Lessons Teaching",
            "type": "array"
          },
          "mastery_distribution": {
            "additionalProperties": true,
            "title": "Mastery Distribution",
            "type": "object"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "category",
          "difficulty",
          "mastery_distribution",
          "incidents_demonstrating",
          "lessons_teaching"
        ],
        "title": "KnowledgeNodeResponse",
        "type": "object"
      },
      "LearningPathResponse": {
        "additionalProperties": false,
        "description": "Ordered concept ids.",
        "properties": {
          "path": {
            "items": {
              "type": "string"
            },
            "title": "Path",
            "type": "array"
          },
          "student_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Student Id"
          },
          "target": {
            "title": "Target",
            "type": "string"
          }
        },
        "required": [
          "target",
          "path"
        ],
        "title": "LearningPathResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "The World's First Agentic AI That Learns by Teaching",
    "title": "TUTORIAL API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/investigations/{incident_id}/chain-of-custody": {
      "get": {
        "description": "Return custody documentation for each evidence item.",
        "operationId": "get_chain_of_custody_api_v1_investigations__incident_id__chain_of_custody_get",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Incident Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustodyReportResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Chain Of Custody",
        "tags": [
          "investigations"
        ]
      }
    },
    "/api/v1/investigations/{incident_id}/evidence": {
      "get": {
        "description": "Return evidence rows for an incident.",
        "operationId": "get_evidence_api_v1_investigations__incident_id__evidence_get",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Incident Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EvidenceResponse"
                  },
                  "title": "Response Get Evidence Api V1 Investigations  Incident Id  Evidence Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Evidence",
        "tags": [
          "investigations"
        ]
      }
    },
    "/api/v1/investigations/{incident_id}/self-corrections": {
      "get": {
        "description": "Return self-correction steps for FIND EVIL! reporting.",
        "operationId": "get_self_corrections_api_v1_investigations__incident_id__self_corrections_get",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Incident Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/InvestigationStepResponse"
                  },
                  "title": "Response Get Self Corrections Api V1 Investigations  Incident Id  Self Corrections Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Self Corrections",
        "tags": [
          "investigations"
        ]
      }
    },
    "/api/v1/investigations/{incident_id}/steps": {
      "get": {
        "description": "Return ordered investigation steps.",
        "operationId": "get_investigation_steps_api_v1_investigations__incident_id__steps_get",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Incident Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/InvestigationStepResponse"
                  },
                  "title": "Response Get Investigation Steps Api V1 Investigations  Incident Id  Steps Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Investigation Steps",
        "tags": [
          "investigations"
        ]
      }
    },
    "/api/v1/knowledge/concepts/{concept_id}": {
      "get": {
        "description": "Return a single concept node.",
        "operationId": "get_concept_api_v1_knowledge_concepts__concept_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "concept_id",
            "required": true,
            "schema": {
              "title": "Concept Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeNodeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Concept",
        "tags": [
          "knowledge"
        ]
      }
    },
    "/api/v1/knowledge/graph": {
      "get": {
        "description": "Return all nodes and edges.",
        "operationId": "get_knowledge_graph_api_v1_knowledge_graph_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeGraphResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Knowledge Graph",
        "tags": [
          "knowledge"
        ]
      }
    },
    "/api/v1/knowledge/learning-path": {
      "get": {
        "description": "Compute prerequisite ordering toward ``target``.",
        "operationId": "get_learning_path_api_v1_knowledge_learning_path_get",
        "parameters": [
          {
            "in": "query",
            "name": "target",
            "required": true,
            "schema": {
              "title": "Target",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "student_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Student Id"
            }
          },
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearningPathResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Learning Path",
        "tags": [
          "knowledge"
        ]
      }
    },
    "/api/v1/knowledge/statistics": {
      "get": {
        "description": "Return node/edge counts.",
        "operationId": "get_knowledge_statistics_api_v1_knowledge_statistics_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "integer"
                  },
                  "title": "Response Get Knowledge Statistics Api V1 Knowledge Statistics Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Knowledge Statistics",
        "tags": [
          "knowledge"
        ]
      }
    },
...
  }
}